refactor!: move and rename queue package files
The queue package files have been moved and renamed in the `justconv`
package. The package name has been changed from `queue` to `justconv`,
and references to the package functions and types have been updated to
reflect this change.

- Files `default_queue.go`, `queue.go`, and `task.go` have been renamed
and moved to the `pkg/justconv` directory.
- Updated import statements and usage within `justconv.go` to use the
new package name.

This change is breaking due to the renaming and restructuring of the
package, which affects all references and imports.